Take qint64's instead of int's in KLimitedIODevice's constructor.
The class attributes were already qint64, and receiving int's made it fail on huge (2GB+) files.

Ack-ed by dfaure.

CCBUG: 216672

svn path=/trunk/KDE/kdelibs/; revision=1187156

+++ kdecore/io/klimitediodevice_p.h
@@ -38,7 +38,7 @@ public:
      * @param start where to start reading (position in bytes)
      * @param length the length of the data to read (in bytes)
      */
-    KLimitedIODevice( QIODevice *dev, int start, int length );
+    KLimitedIODevice( QIODevice *dev, qint64 start, qint64 length );
     virtual ~KLimitedIODevice() {}
 
     virtual bool isSequential() const;
